Move response module to gmp/http
The Response class is http related.
