+++ ng/src/gmp/http/response.js
